+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Unreleased
 
 - `function` module created with `identity`, `compose`, and `flip` functions.
+- The error type of `list.find_map` is now `Nil`.
+- The labels for `list.split` are now `split(list: _, at: _)`.
 
 ## v0.6.0 - 2019-12-23